Appコンポーネントにデータ付与（途中）
