+++ src/pages/find/style.tsx
@@ -7,6 +7,7 @@ import iconSuperLike from 'images/icons/super-like.svg';
 import iconUnLike from 'images/icons/unlike.svg';
 import iconSetting from 'images/icons/setting.svg';
 import iconShop from 'images/icons/shop.svg';
+import iconNew from 'images/icons/new.svg';
 
 const fadeIn = keyframes`
   from {
@@ -128,4 +129,123 @@ export const Shop = styled.div`
   height: 44px;
   margin-left: 14px;
   background: url(${iconShop}) center / 100% 25px no-repeat;
+`;
+
+
+export const Card = styled.div`
+  position: relative;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  margin: 32px 16px 10px 16px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  border-radius: 16px;
+  background: white;
+
+  &::after {
+    content: '';
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 160px;
+    background: linear-gradient(to bottom, rgba(255, 255, 255, 0), rgba(0, 0, 0, 0.57));
+    border-radius: 0 0 16px 16px;
+  }
+`;
+
+export const New = styled.div`
+  position: absolute;
+  top: -6px;
+  right: -6px;
+  width: 44px;
+  height: 44px;
+  z-index: 100;
+  border-radius: 50%;
+  background: ${color.accent} url(${iconNew}) center / 75% no-repeat;
+`;
+
+export const Image = styled.img`
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 16px;
+`;
+
+export const Inner = styled.div`
+  display: flex;
+  align-items: flex-end;
+  z-index: 100;
+`;
+
+export const Profile = styled.div`
+  flex: 0 0 auto;
+  margin-left: 24px;
+  text-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+`;
+
+export const Title = styled.p`
+  margin-bottom: 10px;
+  font-size: 22px;
+  font-weight: bold;
+  color: white;
+`;
+
+export const Text = styled.p`
+  margin-bottom: 18px;
+  font-size: 11px;
+  font-weight: bold;
+  color: white;
+`;
+
+export const ThumbnailList = styled.ul`
+  display: flex;
+  justify-content: flex-end;
+  margin: 0 10px 12px auto;
+  list-style: none;
+`;
+
+interface IThumbnail {
+  isActive: boolean;
+};
+
+export const Thumbnail = styled.li`
+  flex: 0 0 54px;
+  height: 54px;
+  border-radius: 50%;
+  background: white;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  overflow: hidden;
+  box-sizing: border-box;
+  border: solid ${(props: IThumbnail) => props.isActive ? color.accentText + ' 4px' : 'white 2px'};
+  cursor: pointer;
+
+  & + & {
+    margin-left: 4px;
+  }
+
+  > img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+`;
+
+export const Apeal = styled.p`
+  display: flex;
+  align-items: center;
+  height: 66px;
+  padding: 0 12px;
+  margin: 0 12px 12px;
+  background: white;
+  border-radius: 12px;
+  z-index: 100;
+  font-size: 13px;
+  font-weight: bold;
+  color: ${color.text};
+
+  > span > strong {
+    color: ${color.accentText};
+  }
 `;
